Add IO abstractions in preparation for handling redirects
* Low level operations (such as handling UNIX file descriptors or
  Windows HANDLEs) requires nighty features; thus the runtime has been
  gated by a cargo optional feature `runtime`, as well as any direct
  runtime dependencies have been marked as optional as well
* Changed Ast::Redirect variants to hold a u16 instead of a u32
  - A u16 can still address any valid file descriptor (file descriptors
    are defined as an i32), but it avoids holding negative (invalid)
    descriptors

+++ src/syntax/ast/builder.rs
@@ -87,27 +87,27 @@ pub enum WordKind<C> {
 #[derive(Debug)]
 pub enum RedirectKind<W> {
     /// Open a file for reading, e.g. [n]< file
-    Read(Option<u32>, W),
+    Read(Option<u16>, W),
     /// Open a file for writing after truncating, e.g. [n]> file
-    Write(Option<u32>, W),
+    Write(Option<u16>, W),
     /// Open a file for reading and writing, e.g. [n]<> file
-    ReadWrite(Option<u32>, W),
+    ReadWrite(Option<u16>, W),
     /// Open a file for writing, appending to the end, e.g. [n]>> file
-    Append(Option<u32>, W),
+    Append(Option<u16>, W),
     /// Open a file for writing, failing if the `noclobber` shell option is set, e.g.[n]>| file
-    Clobber(Option<u32>, W),
+    Clobber(Option<u16>, W),
     /// Lines contained in the source that should be provided by as input to a file descriptor.
-    Heredoc(Option<u32>, W),
+    Heredoc(Option<u16>, W),
 
     /// Duplicate a file descriptor for reading, e.g. [n]<& n
-    DupRead(Option<u32>, W),
+    DupRead(Option<u16>, W),
     /// Duplicate a file descriptor for writing, e.g. [n]>& n
-    DupWrite(Option<u32>, W),
+    DupWrite(Option<u16>, W),
 
     /// Close a file descriptor for reading, e.g. [n]<&-
-    CloseRead(Option<u32>),
+    CloseRead(Option<u16>),
     /// Close a file descriptor for writing, e.g. [n]>&-
-    CloseWrite(Option<u32>),
+    CloseWrite(Option<u16>),
 }
 
 /// Represents the type of parameter that was parsed
